Refactor na forma de ler os parâmetros

+++ GraphColoring/src/GraphColoring.cpp
@@ -143,7 +143,7 @@ void Problem::writeResultInFile(char *dados, char *parametros, ExecutionInfo inf
 		}
 
 		fprintf(f, "%*d%*d", -16, (int) info.bestFitness, -16, (int) info.worstFitness);
-		fprintf(f, "%*d%*d%*d", -16, info.executionCount, -16, (int) info.executionTime, -24, (int) info.exploredSolutions);
+		fprintf(f, "%*d%*d%*d", -16, (int) info.executionCount, -16, (int) info.executionTime, -24, (int) info.exploredSolutions);
 		fprintf(f, "%*s%s\n", -24, dados, parametros);
 
 		fclose(f);
